requirements.txt
pysdl2
pysdl2-dll
pillow
freetype-py
requests
